fix: design global store properly
